fix more warnings

+++ src/gui/HoverableLabel.h
@@ -89,7 +89,10 @@ private:
         auto x_offset = (getBounds().getWidth() - textWidth) / 2;
         auto y_offset = (getBounds().getHeight() - textHeight) / 2;
 
-        return juce::Rectangle<int>(getX() + x_offset, getY() + y_offset, textWidth, textHeight);
+        return juce::Rectangle<int>(getX() + static_cast<int>(x_offset), 
+                                    getY() + static_cast<int>(y_offset), 
+                                    static_cast<int>(textWidth), 
+                                    static_cast<int>(textHeight));
     }
 
     juce::Colour originalTextColor;
